+++ ICD.Connect.Cameras.Visca/AbstractViscaCameraDeviceSettings.cs
@@ -1,0 +1,201 @@
+﻿using System;
+using ICD.Common.Utils;
+using ICD.Common.Utils.Xml;
+using ICD.Connect.Cameras.Devices;
+using ICD.Connect.Protocol.Ports;
+using ICD.Connect.Protocol.Ports.ComPort;
+using ICD.Connect.Protocol.Settings;
+using ICD.Connect.Settings.Attributes.SettingsProperties;
+
+namespace ICD.Connect.Cameras.Visca
+{
+	public abstract class AbstractViscaCameraDeviceSettings : AbstractCameraDeviceSettings, IViscaCameraDeviceSettings
+	{
+		private const string PORT_ELEMENT = "Port";
+		private const string PAN_TILT_SPEED_ELEMENT = "PanTiltSpeed";
+		private const string ZOOM_SPEED_ELEMENT = "ZoomSpeed";
+
+		private readonly ComSpecProperties m_ComSpecProperties;
+
+		private int? m_PanTiltSpeed;
+		private int? m_ZoomSpeed;
+
+		#region Properties
+
+		[OriginatorIdSettingsProperty(typeof(ISerialPort))]
+		public int? Port { get; set; }
+
+		public int? PanTiltSpeed
+		{
+			get { return m_PanTiltSpeed; }
+			set
+			{
+				if (value == null)
+				{
+					m_PanTiltSpeed = null;
+				}
+				else
+				{
+					m_PanTiltSpeed = MathUtils.Clamp(value.Value, 1, 24);
+				}
+			}
+		}
+
+		public int? ZoomSpeed
+		{
+			get { return m_ZoomSpeed; }
+			set
+			{
+				if (value == null)
+				{
+					m_ZoomSpeed = null;
+				}
+				else
+				{
+					m_ZoomSpeed = MathUtils.Clamp(value.Value, 1, 20);
+				}
+			}
+		}
+
+		#endregion
+
+		#region Com Spec
+
+		/// <summary>
+		/// Gets/sets the configurable baud rate.
+		/// </summary>
+		public eComBaudRates? ComSpecBaudRate
+		{
+			get { return m_ComSpecProperties.ComSpecBaudRate; }
+			set { m_ComSpecProperties.ComSpecBaudRate = value; }
+		}
+
+		/// <summary>
+		/// Gets/sets the configurable number of data bits.
+		/// </summary>
+		public eComDataBits? ComSpecNumberOfDataBits
+		{
+			get { return m_ComSpecProperties.ComSpecNumberOfDataBits; }
+			set { m_ComSpecProperties.ComSpecNumberOfDataBits = value; }
+		}
+
+		/// <summary>
+		/// Gets/sets the configurable parity type.
+		/// </summary>
+		public eComParityType? ComSpecParityType
+		{
+			get { return m_ComSpecProperties.ComSpecParityType; }
+			set { m_ComSpecProperties.ComSpecParityType = value; }
+		}
+
+		/// <summary>
+		/// Gets/sets the configurable number of stop bits.
+		/// </summary>
+		public eComStopBits? ComSpecNumberOfStopBits
+		{
+			get { return m_ComSpecProperties.ComSpecNumberOfStopBits; }
+			set { m_ComSpecProperties.ComSpecNumberOfStopBits = value; }
+		}
+
+		/// <summary>
+		/// Gets/sets the configurable protocol type.
+		/// </summary>
+		public eComProtocolType? ComSpecProtocolType
+		{
+			get { return m_ComSpecProperties.ComSpecProtocolType; }
+			set { m_ComSpecProperties.ComSpecProtocolType = value; }
+		}
+
+		/// <summary>
+		/// Gets/sets the configurable hardware handshake type.
+		/// </summary>
+		public eComHardwareHandshakeType? ComSpecHardwareHandshake
+		{
+			get { return m_ComSpecProperties.ComSpecHardwareHandshake; }
+			set { m_ComSpecProperties.ComSpecHardwareHandshake = value; }
+		}
+
+		/// <summary>
+		/// Gets/sets the configurable software handshake type.
+		/// </summary>
+		public eComSoftwareHandshakeType? ComSpecSoftwareHandshake
+		{
+			get { return m_ComSpecProperties.ComSpecSoftwareHandshake; }
+			set { m_ComSpecProperties.ComSpecSoftwareHandshake = value; }
+		}
+
+		/// <summary>
+		/// Gets/sets the configurable report CTS changes state.
+		/// </summary>
+		public bool? ComSpecReportCtsChanges
+		{
+			get { return m_ComSpecProperties.ComSpecReportCtsChanges; }
+			set { m_ComSpecProperties.ComSpecReportCtsChanges = value; }
+		}
+
+		/// <summary>
+		/// Clears the configured values.
+		/// </summary>
+		void IComSpecProperties.ClearComSpecProperties()
+		{
+			m_ComSpecProperties.ClearComSpecProperties();
+		}
+
+		#endregion
+
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		protected AbstractViscaCameraDeviceSettings()
+		{
+			m_ComSpecProperties = new ComSpecProperties();
+		}
+
+		/// <summary>
+		/// Write settings elements to xml.
+		/// </summary>
+		/// <param name="writer"></param>
+		protected override void WriteElements(IcdXmlTextWriter writer)
+		{
+			base.WriteElements(writer);
+
+			writer.WriteElementString(PORT_ELEMENT, IcdXmlConvert.ToString(Port));
+			writer.WriteElementString(PAN_TILT_SPEED_ELEMENT, IcdXmlConvert.ToString(PanTiltSpeed));
+			writer.WriteElementString(ZOOM_SPEED_ELEMENT, IcdXmlConvert.ToString(ZoomSpeed));
+
+			m_ComSpecProperties.WriteElements(writer);
+		}
+
+		/// <summary>
+		/// Updates the settings from xml.
+		/// </summary>
+		/// <param name="xml"></param>
+		public override void ParseXml(string xml)
+		{
+			base.ParseXml(xml);
+
+			Port = XmlUtils.TryReadChildElementContentAsInt(xml, PORT_ELEMENT);
+			PanTiltSpeed = XmlUtils.TryReadChildElementContentAsInt(xml, PAN_TILT_SPEED_ELEMENT);
+			ZoomSpeed = XmlUtils.TryReadChildElementContentAsInt(xml, ZOOM_SPEED_ELEMENT);
+
+			m_ComSpecProperties.ParseXml(xml);
+
+			UpdateComSpecDefaults(m_ComSpecProperties);
+		}
+
+		private void UpdateComSpecDefaults(ComSpecProperties comSpecProperties)
+		{
+			if(comSpecProperties == null)
+				throw new ArgumentNullException("comSpecProperties");
+
+			comSpecProperties.ApplyDefaultValues(eComBaudRates.BaudRate9600,
+			                                     eComDataBits.DataBits8,
+			                                     eComParityType.None,
+			                                     eComStopBits.StopBits1,
+			                                     eComProtocolType.Rs232,
+			                                     eComHardwareHandshakeType.None,
+			                                     eComSoftwareHandshakeType.None,
+			                                     false);
+		}
+	}
+}