+++ ICD.Connect.Cameras.Visca/Aver/AverPtz3ViscaCameraDeviceSettings.cs
@@ -1,0 +1,9 @@
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Cameras.Visca.Aver
+{
+	[KrangSettings("AverPtz3", typeof(AverPtz3ViscaCameraDevice))]
+	public sealed class AverPtz3ViscaCameraDeviceSettings : AbstractViscaCameraDeviceSettings
+	{
+	}
+}