+++ ICD.Connect.Cameras.Visca/IViscaCameraDevice.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Cameras.Devices;
+using ICD.Connect.Devices;
+
+namespace ICD.Connect.Cameras.Visca
+{
+	public interface IViscaCameraDevice : ICameraDevice, IDeviceWithPower
+	{
+	}
+}